Extraída la lógica de las pistas a un nuevo componente PokemonHints
